Allow clients to accept/reject requests
